refactor: response 대신 request에서 attribute를 설정하도록 수정
