+++ src/main/java/http/request/RequestBody.java
@@ -3,12 +3,14 @@ package http.request;
 import utils.QueryUrlParser;
 
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
 public class RequestBody {
 
     private Map<String, String> body;
+    private Map<String, Object> attributes = new HashMap<>();
 
     private RequestBody(Map<String, String> body) {
         this.body = body;
@@ -24,5 +26,13 @@ public class RequestBody {
 
     public Optional<String> get(String key) {
         return Optional.ofNullable(this.body.get(key));
+    }
+
+    public void addAttribute(String name, Object value) {
+        attributes.put(name, value);
+    }
+
+    public Map<String, Object> getBodyAttributes() {
+        return attributes;
     }
 }
