+++ src/main/java/http/response/ResponseBody.java
@@ -6,7 +6,6 @@ import java.util.Map;
 public class ResponseBody {
 
     private byte[] body;
-    private Map<String, Object> attributes = new HashMap<>();
 
     public ResponseBody(byte[] body) {
         this.body = body;
@@ -18,13 +17,5 @@ public class ResponseBody {
 
     public int getContentLength() {
         return body.length;
-    }
-
-    public void addAttribute(String name, Object value) {
-        attributes.put(name, value);
-    }
-
-    public Map<String, Object> getBodyAttributes() {
-        return attributes;
     }
 }
